refactor

+++ src/main/java/web/controller/UserController.java
@@ -10,6 +10,7 @@ import org.springframework.ui.ModelMap;
 import web.model.*;
 
 import javax.persistence.EntityNotFoundException;
+import java.util.ArrayList;
 import java.util.List;
 
 
@@ -29,14 +30,9 @@ public class UserController {
     @GetMapping(value = "/delete=id{id}")
     public String deleteUser(@PathVariable("id") long id, Model model) {
         User user;
-        if (id >= 0) {
-            try {
-                user = userService.getUserById(id);
+        if (id >= 0 &&  (user = userService.getUserById(id))!=null) {
                 userService.delete(user);
                 return "redirect:users";
-            } catch (EntityNotFoundException e) {
-                e.printStackTrace();
-            }
         }
         model.addAttribute("message", "Ошибка, проверьте ID");
         return "error";
@@ -56,14 +52,9 @@ public class UserController {
     @GetMapping(value = "/edit=id{id}")
     public String editUser(@PathVariable("id") long id, Model model) {
         User user;
-        if (id >= 0) {
-            try {
-                user = userService.getUserById(id);
+        if (id >= 0 &&  (user = userService.getUserById(id))!=null) {
                 model.addAttribute("user", user);
                 return "user-form";
-            } catch (EntityNotFoundException e) {
-                e.printStackTrace();
-            }
         }
         model.addAttribute("message", "Ошибка, проверьте ID");
         return "error";
@@ -73,5 +64,20 @@ public class UserController {
     public String updateUser(@ModelAttribute User user) {
         userService.update(user);
         return "redirect:users";
+    }
+
+    @RequestMapping(value = "hello", method = RequestMethod.GET)
+    public String printWelcome(ModelMap model) {
+        List<String> messages = new ArrayList<>();
+        messages.add("Hello!");
+        messages.add("I'm Spring MVC-SECURITY application");
+        messages.add("5.2.0 version by sep'19 ");
+        model.addAttribute("messages", messages);
+        return "hello";
+    }
+
+    @RequestMapping(value = "login", method = RequestMethod.GET)
+    public String loginPage() {
+        return "login";
     }
 }
